separated Calculator into its own component
