+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/App.js":"1","/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/reportWebVitals.js":"2"},{"size":2378,"mtime":1606433109296,"results":"3","hashOfConfig":"4"},{"size":362,"mtime":1606255637684,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"i019op",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/App.js",[],"/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/reportWebVitals.js",[]]
+[{"/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/App.js":"1","/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/reportWebVitals.js":"2","/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/Calculator.js":"3"},{"size":612,"mtime":1606439601167,"results":"4","hashOfConfig":"5"},{"size":362,"mtime":1606255637684,"results":"6","hashOfConfig":"5"},{"size":1900,"mtime":1606439641624,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"i019op",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/App.js",[],"/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/reportWebVitals.js",[],"/mnt/c/Users/kald3/dev/enzyme-and-graphql-app/src/Calculator.js",[]]
